refactor: remove CallFees re-export and relocate tests
Removed the CallFees re-export from revm_utils.rs and moved all
CallFees-related tests to fees.rs where the struct is defined.
This improves code organization by co-locating tests with their
implementation.
